T3: Added - Made proto file for transaction. Generated grpc classes of transaction service.

+++ SEP3_DataTier/EFCDataAccess/DAOInterface/ITransactionDao.cs
@@ -5,10 +5,10 @@ namespace EFCDataAccess.DAOInterface;
 public interface ITransactionDao
 {
     Task<TransactionEntity> CreateTransactionAsync(TransactionEntity transaction);
-    Task<TransactionEntity?> FetchTransactionByIdAsync(int id);
+    Task<TransactionEntity?> FetchTransactionByIdAsync(long id);
     Task<ICollection<TransactionEntity>> FetchAlLTransactionsBySenderAsync(string senderUsername);
     Task<ICollection<TransactionEntity>> FetchAllTransactionsByReceiverAsync(string receiverUsername);
     Task<ICollection<TransactionEntity>> FetchAlLTransactionsInvolvingUserAsync(string username);
     Task<ICollection<TransactionEntity>> FetchTransactionsByDateAsync(string date);
-    Task<bool> DeleteTransactionAsync(int id);
+    Task<bool> DeleteTransactionAsync(long id);
 }
